change money type from int to double

+++ Library_System/src/model/User.java
@@ -11,7 +11,7 @@ public class User {
     public String type;//本科生（undergraduate），研究生（postgraduate），教师（teacher）
     public int MaxNum;//借书的最大本数
     public int MaxPeriod;//结束的最大时长
-    public int money;//用户的总金额数
+    public double money;//用户的总金额数
 
     public String getId() {
         return id;
@@ -61,11 +61,11 @@ public class User {
         MaxPeriod = maxPeriod;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
 }
